re-organizing calib functions
