+++ calib/calibrate.py
@@ -5,9 +5,10 @@ from pathlib import Path
 import calib_db
 import click
 import optuna
-from calib_report import plot_stuff
-from calib_report import save_study_results
-from logic import run_worker_main
+import sciris as sc
+from report import plot_stuff
+from report import save_study_results
+from worker import run_worker_main
 
 import laser_polio as lp
 
@@ -19,7 +20,7 @@ if os.getenv("POLIO_ROOT"):
 # ------------------- USER CONFIG -------------------
 num_trials = 2
 study_name = "calib_nigeria_smpop_r0_k_seasonality_20250421_date_fix"
-calib_config_path = lp.root / "calib/calib_configs/r0_k_seasonality.yaml"
+calib_config_path = lp.root / "calib/calib_configs/r0_k_ssn.yaml"
 model_config_path = lp.root / "calib/model_configs/config_nigeria_popscale0.01.yaml"
 fit_function = "log_likelihood"  # options are "log_likelihood" or "mse"
 results_path = lp.root / "results" / study_name
@@ -42,7 +43,7 @@ def main(model_config, results_path, study_name, fit_function="mse", **kwargs):
     if not os.getenv("HEADLESS"):
         plot_stuff(study_name, storage_url, output_dir=Path(results_path))
 
-    print("✅ Calibration complete. Results saved.")
+    sc.printcyan("✅ Calibration complete. Results saved.")
 
 
 @click.command(context_settings=CONTEXT_SETTINGS)
